Remove rubocop opt out for Rails/HasManyOrHasOneDependent
Also adds dependent: destroy to ApplicationChoice and ProviderUser

Reason being that not having these was causing the that cop to fire,
but turning of a cop is not a good solution as it may catch something
useful in future.

We also have cascading constraints at the DB level which are mirrored in
these dependent: destroy declarations.