Add to existing providers when the user already exists
This commit adds the selected providers to the providers for an existing user.
This covers the case where a user is not visible to the managing user but is
to be added to the available providers of the managing user.
It comes with the penalty that when updating an existing user this way we cannot
remove any existing providers from them, this would be handled by the edit form.