Show the application that candidates/providers see in support
We currently make support-specific components for certain application
information or using the provider/candidate components directly. This
can cause data leakage, because the support UI will be able to see more
things than the provider UI, and we can easily forget this when
updating components.

This replaces the section in support with an explicit import of the
things that the provider or candidate sees. In addition to achieving
the correct isolation, we make it a lot easier to see whether or not
there are inconsistencies between the 2 UIs. We also don't have to
update the support whenever we update provider or candidate UI.