Consider recruitment cycle when looking up course for UCAS matches
Now that we are approaching the next recruitment cycle we have two courses with
the same code and provider. One for each recruitment cycle year. Decided to
hardcore it instead of using `RecruitmentCycle.current_year`,
`RecruitmentCycle.previous_year` as it’s going to change soon but we still want
to compare courses in 2020 recruitment cycle.